# docs/swagger.yaml
basePath: /
definitions:
...
  models.ConnectionResponse:
    properties:
      createdAt:
        type: string
      credentialsSecret:
        allOf:
        - $ref: '#/definitions/models.CredentialsSecretRef'
        description: |-
          CredentialsSecret is the Secret holding the secret fields, so a consumer
          can reference it from a Deployment without the console ever reading it.
      description:
        type: string
      message:
        type: string
      name:
        type: string
      namespace:
        type: string
      scope:
        type: string
      secretFields:
        items:
          type: string
        type: array
      status:
        type: string
      type:
        type: string
      values:
        additionalProperties: {}
        type: object
    type: object
...
  models.PlatformService:
    properties:
      category:
        type: string
      defaultVersion:
        type: string
      description:
        type: string
      icon:
        type: string
      name:
        type: string
      repository:
        type: string
      versions:
        items:
          type: string
        type: array
    type: object
...
  models.ProfileImage:
    properties:
      image:
        type: string
      label:
        type: string
    type: object
  models.Project:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  service.ServiceVersionsResponse:
    properties:
      default:
        type: string
      versions:
        items:
          type: string
        type: array
    type: object
host: localhost:8093
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Minimal API server for OKDP UI New
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OKDP Server New API
  version: "1.0"
paths:
...
  /api/platform-services:
    get:
      description: Get the list of managed OKDP services that can be deployed per
        project
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PlatformService'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List available platform services
      tags:
      - platform-services
    post:
      consumes:
      - application/json
      description: Expose a new managed service in the catalog (writes the default
        KuboCD Context)
      parameters:
      - description: Catalog service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PlatformService'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PlatformService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a service to the catalog
      tags:
      - platform-services
  /api/platform-services/{serviceName}:
    delete:
      description: Remove an exposed service from the catalog (writes the default
        KuboCD Context)
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a catalog service
      tags:
      - platform-services
    put:
      consumes:
      - application/json
      description: Update an exposed service (versions, default, description, icon,
        category)
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Catalog service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PlatformService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlatformService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a catalog service
      tags:
      - platform-services
  /api/platform-services/{serviceName}/inputs:
    get:
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Package version tag (defaults to Context CR tag)
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PackageInput'
            type: array
      summary: Connections a service's package declares it needs
      tags:
      - platform-services
  /api/platform-services/{serviceName}/schema:
    get:
      description: Returns the JSON Schema (parameters with UI metadata) for a given
        service version
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Package version tag (defaults to Context CR tag)
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get JSON Schema for a platform service
      tags:
      - platform-services
  /api/platform-services/{serviceName}/versions:
    get:
      description: Returns the list of versions declared in the KuboCD Context CR
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ServiceVersionsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List available versions for a platform service
      tags:
      - platform-services
  /api/profile-images:
    get:
      description: Returns the list of available container images per profile type
        from the KuboCD Context
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.ProfileImage'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get available profile images
      tags:
      - platform-services
  /api/projects:
    get:
      consumes:
      - application/json
      description: Get a list of all projects (backed by Kubernetes Namespaces)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project (materialized as a Kubernetes Namespace)
      parameters:
      - description: Project Object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a project
      tags:
      - projects
  /api/projects/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a project by name
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a single project by name
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update a project's mutable metadata (currently its description)
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      - description: Project Object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a project
      tags:
      - projects
  /api/projects/{name}/connections:
    get:
      description: Connections declared in the project namespace. Connections owned
        by a deployed release are excluded; they are returned by the internal endpoint.
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConnectionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the external connections of a project
      tags:
      - connections
    post:
      consumes:
      - application/json
      description: Stores the credential fields in a Kubernetes Secret and creates
        the Connection CRD referencing it
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an external connection in a project
      tags:
      - connections
  /api/projects/{name}/connections/{connName}:
    delete:
      description: Removes the Connection CRD and the Secret holding its credentials
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an external connection of a project
      tags:
      - connections
    put:
      consumes:
      - application/json
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      - description: Connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an external connection of a project
      tags:
      - connections
  /api/projects/{name}/connections/{connName}/consumers:
    get:
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConnectionConsumer'
            type: array
      summary: Services of the project bound to a connection
      tags:
      - connections
  /api/projects/{name}/connections/internal:
    get:
      description: Connections provided by the services already deployed in the project,
        that the project's other services can consume
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InternalConnection'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the internal connections of a project
      tags:
      - connections
  /api/projects/{name}/connections/selectable:
    get:
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Contract the connections must satisfy
        in: query
        name: contract
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SelectableConnection'
            type: array
      summary: Connections offered for a package input of a given contract
      tags:
      - connections
  /api/projects/{name}/connections/test:
    post:
      consumes:
      - application/json
      description: 'Opens a real connection to the endpoint and checks the credentials.
        Nothing is written to the cluster, so a connection can be validated before
        being created. Always answers 200: the outcome is in the body.'
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection settings to probe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConnectionTestResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Test a connection from the platform
      tags:
      - connections
  /api/projects/{name}/services:
    get:
      description: Get all deployed service instances for a project
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ServiceInstance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List deployed services for a project
      tags:
      - services
    post:
      consumes:
      - application/json
      description: Deploy a managed platform service into a project
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service Deploy Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ServiceInstance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deploy a platform service
      tags:
      - services
  /api/projects/{name}/services/{serviceName}:
    delete:
      description: Remove a deployed service from a project
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a deployed service
      tags:
      - services
    get:
      description: Get a specific deployed service instance
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServiceInstance'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a deployed service
      tags:
      - services
...
  /api/projects/{name}/services/{serviceName}/parameters:
    patch:
      consumes:
      - application/json
      description: Merge new parameters and optionally update the package version
        of a deployed service
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service instance name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Update request (tag and/or parameters)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ServiceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServiceInstance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update service parameters and/or version
      tags:
      - services
  /api/projects/{name}/services/{serviceName}/pods:
    get:
      description: Get all pods belonging to a deployed service instance
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service instance name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List pods for a service instance
      tags:
      - services
  /api/projects/{name}/services/{serviceName}/pods/{podName}/logs:
    get:
      description: Retrieve logs for a specific pod. Use follow=true for streaming
        via SSE.
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service instance name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Pod name
        in: path
        name: podName
        required: true
        type: string
      - description: Container name (defaults to first app container)
        in: query
        name: container
        type: string
      - description: Number of lines from the end (default 100)
        in: query
        name: tailLines
        type: integer
      - description: Stream logs in real-time (default false)
        in: query
        name: follow
        type: boolean
      produces:
      - text/plain
      responses:
        "200":
          description: log output
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get pod logs
      tags:
      - services
  /api/projects/{name}/services/stream:
    get:
      description: Stream service status updates using Server-Sent Events (SSE)
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: stream
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream service updates
      tags:
      - services
...
swagger: "2.0"